Implements the Leaderboard calculator
Next up would be the leaderboard that takes the information from the cache and renders each driver's lap time and highlights their fastest lap so far then summarizes the fastest laps at the end of the list

In the future, we'll add roles and permissions and activity logging. We'll also add the capability to add other users to the app.